% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_boin.R
\name{.sim_boin_one_trial}
\alias{.sim_boin_one_trial}
\title{Simulate One BOIN Trial}
\usage{
.sim_boin_one_trial(
  target,
  p_true,
  n_doses,
  n_cohort,
  cohort_size,
  decision_table,
  stopping_boundaries,
  n_earlystop,
  min_mtd_sample,
  cutoff_eli
)
}
\description{
Simulate One BOIN Trial
}
\keyword{internal}
