% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_boin.R
\name{.summarize_simulation}
\alias{.summarize_simulation}
\title{Summarize BOIN Simulation Results}
\usage{
.summarize_simulation(simulation_results, n_doses)
}
\description{
Summarize BOIN Simulation Results
}
\keyword{internal}
